requirements.txt
scapy>=2.5.0
paho-mqtt>=1.6.1
netifaces>=0.11.0
python-nmap>=0.7.1
cryptography>=41.0.0
pyOpenSSL>=23.0.0
colorama>=0.4.6
netaddr>=0.8.0